src/keyvault-kek/requirements.txt
azure-cli==2.1.0
